update get_local_yatai_container function
