+++ tests/yatai/local_yatai_service.py
@@ -29,53 +29,56 @@ def wait_until_container_ready(docker_container, timeout_seconds=120):
 
 
 @contextlib.contextmanager
-def local_yatai_service_container(db_url=None, repo_base_url=None):
+def local_yatai_service_container(image_tag, db_url=None, repo_base_url=None):
     ensure_docker_available_or_raise()
     docker_client = docker.from_env()
-    local_bentoml_repo_path = os.path.abspath(__file__ + "/../../../")
-    yatai_docker_image_tag = f'bentoml/yatai-service:e2e-test-{uuid.uuid4().hex[:6]}'
+    yatai_docker_image_tag = f'bentoml/yatai-service:e2e-test-{image_tag}'
 
-    # Note: When set both `custom_context` and `fileobj`, docker api will not use the
-    #   `path` provide... docker/api/build.py L138. The solution is create an actual
-    #   Dockerfile along with path, instead of fileobj and custom_context.
-    with TempDirectory() as temp_dir:
-        temp_docker_file_path = os.path.join(temp_dir, 'Dockerfile')
-        with open(temp_docker_file_path, 'w') as f:
-            f.write(
-                f"""\
+    try:
+        docker_client.images.get(yatai_docker_image_tag)
+    except docker.errors.ImageNotFound:
+        # Note: When set both `custom_context` and `fileobj`, docker api will not use
+        # the `path` provide... docker/api/build.py L138. The solution is create an
+        # actual Dockerfile along with path, instead of fileobj and custom_context.
+        local_bentoml_repo_path = os.path.abspath(__file__ + "/../../../")
+        with TempDirectory() as temp_dir:
+            temp_docker_file_path = os.path.join(temp_dir, 'Dockerfile')
+            with open(temp_docker_file_path, 'w') as f:
+                f.write(
+                    f"""\
 FROM bentoml/yatai-service:latest
 ADD . /bentoml-local-repo
 RUN pip install -U /bentoml-local-repo
-            """
+                """
+                )
+            logger.info(f'building docker image {yatai_docker_image_tag}')
+            docker_client.images.build(
+                path=local_bentoml_repo_path,
+                dockerfile=temp_docker_file_path,
+                tag=yatai_docker_image_tag,
             )
-        logger.info(f'building docker image {yatai_docker_image_tag}')
-        docker_client.images.build(
-            path=local_bentoml_repo_path,
-            dockerfile=temp_docker_file_path,
-            tag=yatai_docker_image_tag,
-        )
 
-        container_name = f'yatai-service-container-{uuid.uuid4().hex[:6]}'
-        yatai_service_url = 'localhost:50051'
-        yatai_server_command = ['bentoml', 'yatai-service-start', '--no-ui']
-        if db_url:
-            yatai_server_command.extend(['--db-url', db_url])
-        if repo_base_url:
-            yatai_server_command.extend(['--repo-base-url', repo_base_url])
-        container = docker_client.containers.run(
-            image=yatai_docker_image_tag,
-            environment=['BENTOML_HOME=/tmp'],
-            ports={'50051/tcp': 50051},
-            command=yatai_server_command,
-            name=container_name,
-            detach=True,
-        )
+    container_name = f'yatai-service-container-{uuid.uuid4().hex[:6]}'
+    yatai_service_url = 'localhost:50051'
+    yatai_server_command = ['bentoml', 'yatai-service-start', '--no-ui']
+    if db_url:
+        yatai_server_command.extend(['--db-url', db_url])
+    if repo_base_url:
+        yatai_server_command.extend(['--repo-base-url', repo_base_url])
+    container = docker_client.containers.run(
+        image=yatai_docker_image_tag,
+        environment=['BENTOML_HOME=/tmp'],
+        ports={'50051/tcp': 50051},
+        command=yatai_server_command,
+        name=container_name,
+        detach=True,
+    )
 
-        wait_until_container_ready(container)
-        yield yatai_service_url
+    wait_until_container_ready(container)
+    yield yatai_service_url
 
-        logger.info(f"Shutting down docker container: {container_name}")
-        container.kill()
+    logger.info(f"Shutting down docker container: {container_name}")
+    container.kill()
 
 
 @contextlib.contextmanager
